+++ src/-private/live-query.ts
@@ -6,8 +6,8 @@ import {
 import { notifyPropertyChange } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
 import { createCache, getValue } from '@glimmer/tracking/primitives/cache';
-import { Model } from '../index.ts';
 import type Cache from './cache.ts';
+import type Model from './model.ts';
 import { Orbit } from '@orbit/core';
 import type { SyncLiveQuery } from '@orbit/record-cache';
 import type { RecordQuery, RecordQueryResult } from '@orbit/records';
