+++ src/-private/utils/property-cache.ts
@@ -2,7 +2,7 @@ import { notifyPropertyChange as emberNotifyPropertyChange } from '@ember/object
 import { DEBUG } from '@glimmer/env';
 import { tracked } from '@glimmer/tracking';
 import { createCache, getValue } from '@glimmer/tracking/primitives/cache';
-import Model from '../model.ts';
+import type Model from '../model.ts';
 import { Orbit } from '@orbit/core';
 
 type Cache<T> = ReturnType<typeof createCache<T>>;
